chore: Fix some "use of a fallible conversion" warnings
Fixes the following warnings:

    warning: use of a fallible conversion when an infallible one could be used
       --> src/anlz.rs:635:18
        |
    635 |     #[br(count = len_cues)]
        |                  ^^^^^^^^ help: use: `From::from`
        |
        = note: converting `u16` to `usize` cannot fail
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_fallible_conversions
    note: the lint level is defined here
       --> src/lib.rs:18:9
        |
    18  | #![deny(clippy::all)]
        |         ^^^^^^^^^^^
        = note: `#[warn(clippy::unnecessary_fallible_conversions)]` implied by `#[warn(clippy::all)]`

    warning: use of a fallible conversion when an infallible one could be used
       --> src/anlz.rs:656:18
        |
    656 |     #[br(count = len_cues)]
        |                  ^^^^^^^^ help: use: `From::from`
        |
        = note: converting `u16` to `usize` cannot fail
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_fallible_conversions

    warning: use of a fallible conversion when an infallible one could be used
       --> src/anlz.rs:853:18
        |
    853 |     #[br(count = len_entries)]
        |                  ^^^^^^^^^^^ help: use: `From::from`
        |
        = note: converting `u16` to `usize` cannot fail
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_fallible_conversions

    warning: use of a fallible conversion when an infallible one could be used
       --> src/pdb/string.rs:164:22
        |
    164 |         #[br(count = (header >> 1) - 1)]
        |                      ^^^^^^^^^^^^^ help: use: `From::from`
        |
        = note: converting `u8` to `usize` cannot fail
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_fallible_conversions

    warning: use of a fallible conversion when an infallible one could be used
       --> src/pdb/string.rs:196:24
        |
    196 |     Ascii(#[br(count = len)] Vec<u8>),
        |                        ^^^ help: use: `From::from`
        |
        = note: converting `u16` to `usize` cannot fail
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_fallible_conversions

    warning: use of a fallible conversion when an infallible one could be used
       --> src/pdb/string.rs:199:25
        |
    199 |     Ucs2le(#[br(count = len / 2)] Vec<u16>),
        |                         ^^^ help: use: `From::from`
        |
        = note: converting `u16` to `usize` cannot fail
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_fallible_conversions

    warning: `rekordcrate` (lib) generated 6 warnings
    warning: `rekordcrate` (lib test) generated 6 warnings (6 duplicates)

+++ src/pdb/string.rs
@@ -161,7 +161,7 @@ enum DeviceSQLStringImpl {
         #[bw(calc = (((content.len() + 1) << 1) | 1) as u8)]
         header: u8,
 
-        #[br(count = (header >> 1) - 1)]
+        #[br(count = usize::from(header >> 1) - 1)]
         content: Vec<u8>,
     },
     /// Regular long form strings, containing possibly different encodings
@@ -193,10 +193,10 @@ enum LongBody {
     #[br(pre_assert(flags == 0x90))]
     Isrc(#[brw(magic = 0x3u8)] binrw::NullString),
     #[br(pre_assert(flags == 0x40))]
-    Ascii(#[br(count = len)] Vec<u8>),
+    Ascii(#[br(count = usize::from(len))] Vec<u8>),
     #[br(pre_assert(flags == 0x90))]
     #[br(pre_assert(len % 2 == 0))]
-    Ucs2le(#[br(count = len / 2)] Vec<u16>),
+    Ucs2le(#[br(count = usize::from(len / 2))] Vec<u16>),
 }
 
 impl LongBody {
